health-ml-backend/requirements.txt
absl-py==2.1.0
annotated-types==0.6.0
anyio==4.3.0
blinker==1.7.0
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.1
cycler==0.12.1
exceptiongroup==1.2.1
fastapi==0.110.0
filelock==3.13.3
flask==3.0.2
flask-cors==4.0.0
fonttools==4.51.0
gunicorn==20.1.0
google-api-core==2.18.0
google-api-python-client==2.126.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.4
googleapis-common-protos==1.69.2
grpcio==1.51.1
grpcio-status==1.51.1
httplib2==0.22.0
idna==3.10
ipykernel==6.29.5
ipython==9.0.2
ipython-pygments-lexers==1.1.1
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.4.2
jupyter-client==8.6.3
jupyter-core==5.7.2
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.1
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
numpy>=1.23.2,<1.28.0
packaging==24.2
pandas==2.2.3
parso==0.8.4
pickle-mixin==1.0.2
pillow==11.1.0
platformdirs==4.3.7
prompt-toolkit==3.0.50
proto-plus==1.26.1
psutil==7.0.0
pyasn1==0.6.1
pydantic==2.11.1
pydantic-core==2.33.0
Pygments==2.19.1
pymongo==4.11.3
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.1
scipy==1.11.3
scikit-learn==1.5.1
six>=1.5
tornado==6.5
tzdata>=2022.7
wcwidth==0.2.13
